# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI AgentLab API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CompareRequest {
  prompt: string;
  systemPrompt?: string;
  providers: string[];
  temperature?: number;
}

export interface ProviderResult {
  provider: string;
  model: string;
  text: string;
  latencyMs: number;
  estimatedCost: number;
  tokenCount: number;
  qualityScore: number;
  clarityScore: number;
  toneScore: number;
  overallScore: number;
  isDemo: boolean;
}

export interface CompareResponse {
  results: ProviderResult[];
  recommendedWinner: string;
  bestQuality: string;
  cheapest: string;
  fastest: string;
}

export interface HistoryItem {
  id: number;
  name: string;
  prompt: string;
  systemPrompt?: string;
  providers: string[];
  winner: string;
  temperature: number;
  results: string;
  createdAt: string;
}

export interface HistoryListResponse {
  items: HistoryItem[];
}

export interface SaveRunRequest {
  name: string;
  prompt: string;
  systemPrompt?: string;
  providers: string[];
  winner: string;
  temperature: number;
  results: string;
}

export interface DeleteResponse {
  success: boolean;
}

export interface ProviderStatus {
  provider: string;
  connected: boolean;
  model: string;
  description: string;
}

export interface SettingsResponse {
  providers: ProviderStatus[];
}

export interface SaveApiKeyRequest {
  provider: string;
  apiKey: string;
}

export interface SaveApiKeyResponse {
  success: boolean;
  message: string;
}

export interface TestConnectionRequest {
  provider: string;
}

export interface TestConnectionResponse {
  success: boolean;
  message: string;
  latencyMs: number;
}