# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI AgentLab API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Run a prompt comparison across multiple AI providers
 * @summary Run AI comparison
 */
export const RunComparisonBody = zod.object({
  prompt: zod.string(),
  systemPrompt: zod.string().optional(),
  providers: zod.array(zod.string()),
  temperature: zod.number().optional(),
});

export const RunComparisonResponse = zod.object({
  results: zod.array(
    zod.object({
      provider: zod.string(),
      model: zod.string(),
      text: zod.string(),
      latencyMs: zod.number(),
      estimatedCost: zod.number(),
      tokenCount: zod.number(),
      qualityScore: zod.number(),
      clarityScore: zod.number(),
      toneScore: zod.number(),
      overallScore: zod.number(),
      isDemo: zod.boolean(),
    }),
  ),
  recommendedWinner: zod.string(),
  bestQuality: zod.string(),
  cheapest: zod.string(),
  fastest: zod.string(),
});

/**
 * Returns all saved test runs
 * @summary Get test run history
 */
export const GetHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      prompt: zod.string(),
      systemPrompt: zod.string().optional(),
      providers: zod.array(zod.string()),
      winner: zod.string(),
      temperature: zod.number(),
      results: zod.string(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * Save a comparison run to history
 * @summary Save a test run
 */
export const SaveRunBody = zod.object({
  name: zod.string(),
  prompt: zod.string(),
  systemPrompt: zod.string().optional(),
  providers: zod.array(zod.string()),
  winner: zod.string(),
  temperature: zod.number(),
  results: zod.string(),
});

export const SaveRunResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  prompt: zod.string(),
  systemPrompt: zod.string().optional(),
  providers: zod.array(zod.string()),
  winner: zod.string(),
  temperature: zod.number(),
  results: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get a specific run
 */
export const GetRunParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRunResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  prompt: zod.string(),
  systemPrompt: zod.string().optional(),
  providers: zod.array(zod.string()),
  winner: zod.string(),
  temperature: zod.number(),
  results: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a test run
 */
export const DeleteRunParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteRunResponse = zod.object({
  success: zod.boolean(),
});

/**
 * Returns which providers have API keys configured
 * @summary Get provider connection statuses
 */
export const GetSettingsResponse = zod.object({
  providers: zod.array(
    zod.object({
      provider: zod.string(),
      connected: zod.boolean(),
      model: zod.string(),
      description: zod.string(),
    }),
  ),
});

/**
 * @summary Save an API key for a provider
 */
export const SaveApiKeyBody = zod.object({
  provider: zod.string(),
  apiKey: zod.string(),
});

export const SaveApiKeyResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Test a provider connection
 */
export const TestConnectionBody = zod.object({
  provider: zod.string(),
});

export const TestConnectionResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  latencyMs: zod.number(),
});