# lib/api-zod/src/generated/types/historyListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI AgentLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryItem } from "./historyItem";

export interface HistoryListResponse {
  items: HistoryItem[];
}