# lib/api-zod/src/generated/types/providerResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI AgentLab API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderResult {
  provider: string;
  model: string;
  text: string;
  latencyMs: number;
  estimatedCost: number;
  tokenCount: number;
  qualityScore: number;
  clarityScore: number;
  toneScore: number;
  overallScore: number;
  isDemo: boolean;
}