# lib/api-zod/src/generated/types/saveRunRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI AgentLab API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveRunRequest {
  name: string;
  prompt: string;
  systemPrompt?: string;
  providers: string[];
  winner: string;
  temperature: number;
  results: string;
}