# lib/api-zod/src/generated/types/testConnectionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI AgentLab API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TestConnectionResponse {
  success: boolean;
  message: string;
  latencyMs: number;
}